create middleware is-auth.js

+++ src/graphql/resolvers/booking.js
@@ -13,23 +13,40 @@ module.exports = {
       throw err
     }
   },
-  bookEvent: async args => {
+  bookEvent: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!')
+    }
+
     const fetchedEvent = await Event.findOne({ _id: args.eventId })
     const booking = new Booking({
-      user: '5c777b827d44f350bcb48b43',
-      event: fetchedEvent
+      user: req.userId,
+      event: fetchedEvent,
     })
     const result = await booking.save()
     return transformBooking(result)
   },
-  cancelBooking: async args => {
+  cancelBooking: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated!')
+    }
+
     try {
       const booking = await Booking.findById(args.bookingId).populate('event')
+
+      if (!booking) {
+        throw new Error('Booking not found.')
+      }
+
+      if (req.userId !== booking.user.toString()) {
+        throw new Error('No Permission!')
+      }
+
       const event = transformEvent(booking.event)
       await Booking.deleteOne({ _id: args.bookingId })
       return event
     } catch (err) {
       throw err
     }
-  }
+  },
 }
